+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Added
+- Automated annotation of SNVs and small indels.
+    - Disable with `--annotation false`
+
 ### Changed
 - `Input options` and `Output options` have been combined in the `Main options` category